requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
idna==3.4
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0
qrcode==7.4.2
pytest==7.4.3
requests
# sqlite3 is part of the Python standard library – no extra package needed